MyUsaCandidateBackEnd/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
distlib==0.3.9
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
filelock==3.18.0
packaging==25.0
pipenv==2025.0.3
platformdirs==4.3.8
setuptools==80.9.0
sqlparse==0.5.3
virtualenv==20.31.2
